global: renaming core to ext and pytest-runner
* Updates package structure with pytest-runner and renames core to ext.
