+++ run-tests.sh
@@ -26,5 +26,5 @@ pydocstyle invenio_db && \
 isort -rc -c -df **/*.py && \
 check-manifest --ignore ".travis-*" && \
 sphinx-build -qnNW docs docs/_build/html && \
-python setup.py test -a'-vv' && \
+python setup.py test && \
 sphinx-build -qnNW -b doctest docs docs/_build/doctest
